- Started to implement DigitialIO plugin for dynamic clamp
git-svn-id: https://svn.code.sf.net/p/relacs/code/relacs/trunk@780 1550813e-0d35-0410-aec7-fc86b090456a

+++ plugins/linuxdevices/comedi/src/dynclampanaloginput.cc
@@ -749,7 +749,7 @@ int DynClampAnalogInput::convertData( void )
   float *db = (float *)Buffer;
 
   // debug:
-  if ( BufferN*BufferElemSize > BufferSize )
+  if ( BufferN*(int)BufferElemSize > BufferSize )
     cerr << "DynClampAnalogInput::convertData: buffer overflow! BufferN=" << BufferN
 	 << " BufferSize=" << BufferSize << " BufferN*BufferElemSize=" << BufferN*BufferElemSize << '\n';
